+++ package.json
@@ -22,7 +22,8 @@
     "js-base64": "^3.7.5",
     "typed.js": "^2.0.16",
     "vue": "^3.3.4",
-    "vue-head": "^2.2.0"
+    "vue-head": "^2.2.0",
+    "vue-router": "^4.2.5"
   },
   "devDependencies": {
     "@vitejs/plugin-vue": "^4.4.1",
